--- frontend/src/features/campaign/routes/index.tsx
@@ -1,1 +1,0 @@
-export * from './Campaign';
