+++ frontend/src/features/mint/routes/index.tsx
@@ -1,0 +1,1 @@
+export * from './Mint';
