+++ frontend/src/routes/public.tsx
@@ -3,9 +3,9 @@ import { Outlet } from 'react-router-dom';
 
 import { Spinner } from '@/components/Elements';
 import { MainLayout } from '@/components/Layout';
-import { Campaign } from '@/features/campaign';
 import { Campaigns } from '@/features/campaigns';
 import { Donate } from '@/features/donate';
+import { Mint } from '@/features/mint';
 import { Landing } from '@/features/misc';
 
 const App = () => {
@@ -32,7 +32,7 @@ export const publicRoutes = [
       { path: '/', element: <Landing /> },
       { path: '/donate/:id', element: <Donate /> },
       { path: '/campaigns', element: <Campaigns /> },
-      { path: '/campaign', element: <Campaign /> },
+      { path: '/mint', element: <Mint /> },
     ],
   },
 ];
